requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
xgboost>=1.7.0
plotly>=5.13.0
streamlit>=1.22.0
pyyaml>=6.0
openpyxl>=3.0.0
networkx>=3.0
jupyter>=1.0.0